product and some validation fixed

+++ RR-Commerce/src/modules/product/product.service.ts
@@ -1,8 +1,13 @@
+import type { NextFunction, Request } from "express";
 import type { IProduct } from "./product.interface.js";
 import { Product } from "./product.model.js";
+import User from "../auth/user.model.js";
+import createHttpError from "http-errors";
 
-const Add = async (payload: IProduct) => {
-  return await Product.create(payload);
+const Add = async (req: Request, payload: IProduct) => {
+  const id = req.vendorId;
+
+  return await Product.create({ ...payload, seller: id });
 };
 const Get = async () => {
   return await Product.find();
@@ -10,11 +15,35 @@ const Get = async () => {
 const GetOne = async (id: string) => {
   return await Product.findById(id);
 };
-const Update = async (id: string, payload: IProduct) => {
+const Update = async (
+  req: Request,
+  id: string,
+  payload: IProduct,
+  next: NextFunction
+) => {
+  const vendorId = req.vendorId;
+  const product = await Product.findOne({ _id: id, seller: vendorId });
+
+  if (!product) {
+    return next(
+      createHttpError(404, "Product Not Found or vendor didn't matched")
+    );
+  }
+
   return await Product.findByIdAndUpdate(id, payload, { new: true });
 };
 
-const Delete = async (id: string) => {
+const Delete = async (req: Request, id: string, next: NextFunction) => {
+  const vendorId = req.vendorId;
+
+  const product = await Product.findOne({ _id: id, seller: vendorId });
+
+  if (!product) {
+    return next(
+      createHttpError(404, "Product Not Found or vendor didn't matched")
+    );
+  }
+
   return await Product.findByIdAndDelete(id);
 };
